+++ J81.ADCToolkit/Public/Connect-ADCNode.ps1
@@ -10,6 +10,8 @@ function Connect-ADCNode {
         The credential to authenticate to the ADC with.
     .PARAMETER Timeout
         Timeout in seconds for session object.
+    .PARAMETER HA
+        Connect all possible HA Nodes
     .PARAMETER PassThru
         Return the ADC session object.
     .EXAMPLE
@@ -26,15 +28,21 @@ function Connect-ADCNode {
         [System.Management.Automation.Credential()]$Credential = (Get-Credential -Message "Enter username and password for the Citrix ADC`r`nE.g. nsroot / P@ssw0rd" ),
         
         [int]$Timeout = 900,
-    
+
+        [switch]$HA,
+
         [switch]$PassThru
     )
     # Based on https://github.com/devblackops/NetScaler
     
-    if ($ManagementURL.scheme -eq "https") {
-        Write-Verbose "Connecting to $ManagementURL... Connection is SSL, Ignoring SSL checks"
-        if (-Not ("TrustAllCertsPolicy" -as [type])) {
-            Add-Type -TypeDefinition @"
+    if ($HA) {
+        $ADCSession = Connect-ADCHANodes -ManagementURL $ManagementURL -Credential $Credential -PassThru
+    } else {
+
+        if ($ManagementURL.scheme -eq "https") {
+            Write-Verbose "Connecting to $ManagementURL... Connection is SSL, Ignoring SSL checks"
+            if (-Not ("TrustAllCertsPolicy" -as [type])) {
+                Add-Type -TypeDefinition @"
 using System.Net;
 using System.Security.Cryptography.X509Certificates;
 public class TrustAllCertsPolicy : ICertificatePolicy {
@@ -45,67 +53,68 @@ public class TrustAllCertsPolicy : ICertificatePolicy {
     }
 }
 "@ 
-        }
-        [System.Net.ServicePointManager]::CertificatePolicy = New-Object TrustAllCertsPolicy
-        # Set Tls versions
-        [System.Net.ServicePointManager]::SecurityProtocol = 
-        [System.Net.SecurityProtocolType]::Tls13 -bor `
-            [System.Net.SecurityProtocolType]::Tls12 -bor `
-            [System.Net.SecurityProtocolType]::Tls11
-    } else {
-        Write-Verbose -Message "Connecting to $ManagementURL..."
-    }
-    try {
-        $login = @{
-            login = @{
-                username = $Credential.UserName;
-                password = $Credential.GetNetworkCredential().Password
-                timeout  = $Timeout
             }
-        }
-        $loginJson = ConvertTo-Json -InputObject $login
-        $saveSession = @{ }
-        $params = @{
-            Uri             = "$($ManagementURL.AbsoluteUri)nitro/v1/config/login"
-            Method          = 'POST'
-            Body            = $loginJson
-            SessionVariable = 'saveSession'
-            ContentType     = 'application/json'
-            ErrorVariable   = 'restError'
-            Verbose         = $false
-        }
-        $response = Invoke-RestMethod @params
-    
-        if ($response.severity -eq 'ERROR') {
-            throw "Error. See response: `n$($response | Format-List -Property * | Out-String)"
+            [System.Net.ServicePointManager]::CertificatePolicy = New-Object TrustAllCertsPolicy
+            # Set Tls versions
+            [System.Net.ServicePointManager]::SecurityProtocol = 
+            [System.Net.SecurityProtocolType]::Tls13 -bor `
+                [System.Net.SecurityProtocolType]::Tls12 -bor `
+                [System.Net.SecurityProtocolType]::Tls11
         } else {
-            Write-Verbose -Message "Response:`n$(ConvertTo-Json -InputObject $response | Out-String)"
+            Write-Verbose -Message "Connecting to $ManagementURL..."
         }
-    } catch [Exception] {
-        throw $_
-    }
-    $ADCSession = [PSObject]@{
-        ManagementURL = $ManagementURL;
-        WebSession    = [Microsoft.PowerShell.Commands.WebRequestSession]$saveSession;
-        Username      = $Credential.UserName;
-        Version       = "UNKNOWN";
-    }
+        try {
+            $login = @{
+                login = @{
+                    username = $Credential.UserName;
+                    password = $Credential.GetNetworkCredential().Password
+                    timeout  = $Timeout
+                }
+            }
+            $loginJson = ConvertTo-Json -InputObject $login
+            $saveSession = @{ }
+            $params = @{
+                Uri             = "$($ManagementURL.AbsoluteUri)nitro/v1/config/login"
+                Method          = 'POST'
+                Body            = $loginJson
+                SessionVariable = 'saveSession'
+                ContentType     = 'application/json'
+                ErrorVariable   = 'restError'
+                Verbose         = $false
+            }
+            $response = Invoke-RestMethod @params
     
-    try {
-        Write-Verbose -Message "Trying to retrieve the ADC version"
-        $params = @{
-            Uri           = "$($ManagementURL.AbsoluteUri)nitro/v1/config/nsversion"
-            Method        = 'GET'
-            WebSession    = $ADCSession.WebSession
-            ContentType   = 'application/json'
-            ErrorVariable = 'restError'
-            Verbose       = $false
+            if ($response.severity -eq 'ERROR') {
+                throw "Error. See response: `n$($response | Format-List -Property * | Out-String)"
+            } else {
+                Write-Verbose -Message "Response:`n$(ConvertTo-Json -InputObject $response | Out-String)"
+            }
+        } catch [Exception] {
+            throw $_
         }
-        $response = Invoke-RestMethod @params
-        Write-Verbose -Message "Response:`n$(ConvertTo-Json -InputObject $response | Out-String)"
-        $ADCSession.version = ($response.nsversion.version.Split(","))[0]
-    } catch {
-        Write-Verbose -Message "Error. See response: `n$($response | Format-List -Property * | Out-String)"
+        $ADCSession = [PSObject]@{
+            ManagementURL = $ManagementURL;
+            WebSession    = [Microsoft.PowerShell.Commands.WebRequestSession]$saveSession;
+            Username      = $Credential.UserName;
+            Version       = "UNKNOWN";
+        }
+    
+        try {
+            Write-Verbose -Message "Trying to retrieve the ADC version"
+            $params = @{
+                Uri           = "$($ManagementURL.AbsoluteUri)nitro/v1/config/nsversion"
+                Method        = 'GET'
+                WebSession    = $ADCSession.WebSession
+                ContentType   = 'application/json'
+                ErrorVariable = 'restError'
+                Verbose       = $false
+            }
+            $response = Invoke-RestMethod @params
+            Write-Verbose -Message "Response:`n$(ConvertTo-Json -InputObject $response | Out-String)"
+            $ADCSession.version = ($response.nsversion.version.Split(","))[0]
+        } catch {
+            Write-Verbose -Message "Error. See response: `n$($response | Format-List -Property * | Out-String)"
+        }
     }
     $Script:ADCSession = $ADCSession